requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
django-cleanup==8.0.0
django-imagekit==5.0.0
python-decouple==3.8
whitenoise==6.6.0
django-cors-headers==4.3.1
django-extensions==3.2.3
dj-database-url==2.1.0 